+++ interfaces/update_interfaces.go
@@ -1,6 +1,6 @@
 package interfaces
 
 type IUpdateInventory interface {
-	UpdatedInventory(sku string, soldQuantity float64) error
+	UpdatedInventory(sku string, soldQuantity float32) error
 	IsInStock(sku string) (bool, error)
 }
